Brandon Le: Fixing backend and frontend communication when there is no match so that the frontend can respond appropriately. Reorganizing file structure since there were issues with node modules for frontend.
